updated accelerate examples
