--- examples/big_model_offloading/big_model_fp8.py
@@ -1,35 +1,0 @@
-from transformers import AutoTokenizer
-
-from llmcompressor.modifiers.quantization import QuantizationModifier
-from llmcompressor.transformers import SparseAutoModelForCausalLM, oneshot
-from llmcompressor.transformers.compression.helpers import (  # noqa
-    calculate_offload_device_map,
-)
-
-from llmcompressor.modifiers.quantization import QuantizationModifier
-from llmcompressor.transformers import SparseAutoModelForCausalLM, oneshot
-from transformers import AutoTokenizer
-
-MODEL_ID = "meta-llama/Meta-Llama-3-70B-Instruct"
-
-# Load model.
-model = SparseAutoModelForCausalLM.from_pretrained(
-    MODEL_ID, device_map="auto", torch_dtype="auto")
-tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
-
-# Configure the quantization algorithm and scheme.
-# In this case, we:
-#   * quantize the weights to fp8 with per channel with ptq
-#   * quantize the activations to fp8 dynamic per token
-recipe = QuantizationModifier(targets="Linear", 
-                              scheme="FP8_Dynamic", 
-                              ignore=["lm_head"])
-
-# Apply quantization.
-oneshot(model=model, recipe=recipe)
-
-# Save to disk in compressed-tensors format.
-SAVE_DIR = MODEL_ID.split("/")[1] + "-FP8-Dynamic"
-model.save_pretrained(SAVE_DIR)
-tokenizer.save_pretrained(SAVE_DIR)
-
