+++ examples/big_models_with_accelerate/cpu_offloading_fp8.py
@@ -1,0 +1,21 @@
+from transformers import AutoTokenizer
+from llmcompressor.modifiers.quantization import QuantizationModifier
+from llmcompressor.transformers import SparseAutoModelForCausalLM, oneshot
+
+MODEL_ID = "meta-llama/Meta-Llama-3-70B-Instruct"
+OUTPUT_DIR = MODEL_ID.split("/")[1] + "-FP8-Dynamic"
+
+# Load model
+# Note: device_map="auto" will offload to CPU if not enough space on GPU.
+model = SparseAutoModelForCausalLM.from_pretrained(
+    MODEL_ID, device_map="auto", torch_dtype="auto")
+
+# Configure the quantization scheme and algorithm (PTQ + FP8_DYNAMIC).
+recipe = QuantizationModifier(
+    targets="Linear", scheme="FP8_DYNAMIC", ignore=["lm_head"])
+
+# Apply quantization and save in `compressed-tensors` format.
+oneshot(model=model,
+        recipe=recipe,
+        tokenizer=AutoTokenizer.from_pretrained(MODEL_ID), 
+        output_dir=OUTPUT_DIR)
